Move HC entrypoint into its own `lib` directory
-  Dep cruiser rules updated accordingly

JIRA: RAIL-1815
