+++ libs/sdk-ui-charts/.dependency-cruiser.js
@@ -7,6 +7,8 @@ options = {
         depCruiser.isolatedSubmodule("chart-interfaces", "src/interfaces"),
         depCruiser.moduleWithDependencies("highcharts", "src/highcharts", ["src/interfaces"]),
         depCruiser.moduleWithDependencies("charts", "src/charts", ["src/highcharts", "src/interfaces"]),
+        /* These appear as desired deps for the two modules; currently there are validation errors. The
+         * refactoring should by driven by need to remove these validation errors.
         depCruiser.moduleWithDependencies(
             "chart-options-builder",
             "src/highcharts/chartTypes/_chartOptions",
@@ -15,6 +17,7 @@ options = {
                 "src/interfaces",
                 "src/highcharts/typings/*",
                 "src/highcharts/constants/*",
+                "src/highcharts/lib",
             ],
         ),
         depCruiser.moduleWithDependencies("chart-creators", "src/highcharts/chartTypes/_chartCreators", [
@@ -22,7 +25,8 @@ options = {
             "src/interfaces",
             "src/highcharts/typings/*",
             "src/highcharts/constants/*",
-        ]),
+            "src/highcharts/lib",
+        ]),*/
     ],
     options: depCruiser.DefaultOptions,
 };
