+++ libs/sdk-ui-charts/src/highcharts/chartTypes/heatmap/heatmapConfiguration.ts
@@ -1,7 +1,7 @@
 // (C) 2007-2020 GoodData Corporation
 import cloneDeep from "lodash/cloneDeep";
 import last from "lodash/last";
-import Highcharts from "../../adapter/highcharts";
+import Highcharts from "../../lib";
 import {
     IHighchartsAxisExtend,
     IHighchartsSeriesExtend,
