+++ libs/sdk-ui-charts/src/highcharts/lib/index.ts
